+++ src/io/hdf5/vec_klgs_2010_io.cpp
@@ -20,6 +20,7 @@ VecKlgs2010Input read_input(const ModelFile &file)
     read_mat_if_present(file, "/data/train/w", input.train.w);
     read_mat_if_present(file, "/data/train/x", input.train.x);
     read_forecast_regressors(file, input.spec.k, input.forecast.x);
+    read_test_observations(file, input.spec, input.test.y);
 
     if (input.use_a())
     {